+++ CHANGES.rst
@@ -16,8 +16,10 @@ Internal changes
 
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-* Standardized indices (``xclim.indices.standardized_precipitation_index`` and ``xclim.indices.standardized_precipitation_evapotranspiration_index``) are faster. (:issue:`1270` ,:pull:`1311`).
-* Standardized indicators can be computed in two steps: First compute fit parameters with ``xclim.indices.standardized_index_fit_params``, then use the output in the standardized indices functions.  (:issue:`1270` ,:pull:`1311`).
+* Standardized indices (``xclim.indices.standardized_precipitation_index`` and ``xclim.indices.standardized_precipitation_evapotranspiration_index``)  (:issue:`1270`, :issue:`1416`, ,:pull:`1311`) were changed:
+    * More optimized
+    * Can be computed in two steps: First compute fit parameters with ``xclim.indices.standardized_index_fit_params``, then use the output in the standardized indices functions.
+    * The standardized index values are now clipped by ± 8.21.  This reflects the float64 precision of the computation when cdf values are inverted to a normal distribution.
 
 Breaking changes
 ^^^^^^^^^^^^^^^^